make the skeleton of the game, make branch, talk with group member
Map
The map size is 10x10 (to be determined), with a starting point and an endpoint (the princess). The strength of monsters varies in different locations (different health points). As the game progresses, more powerful weapons will appear. Monsters and bonuses are initially invisible and will only become visible when the player explores the maze and gets within a certain range.

Enemy
Enemies have dialogues, and before combat, they can ask the player whether they want to engage in a battle or retreat for now (this applies to enemies not on the main path). The player does not know the enemy's health, so battles are fought blindly.

Inventory System
Stores HP recovery items.
Items can be discarded.
Sources:
"?" boxes on the ground that can be opened to get random items.
Not obtained from defeating monsters, as that interaction is too complicated.
Loot (Bonus)
Can be either weapons or HP recovery items.
Also found on the ground in "?" boxes.
Not obtained from defeating monsters to avoid complex interactions.
Attack
Use weapons to attack (HP-Attack).
Revival
The player has multiple lives. Upon defeat, the player will be revived with the NPC layout from the current level right before death, allowing them to attack again.
If all lives are used up, the game restarts from the beginning.
Color
Use escape characters to change text color (not implemented yet).